+++ custom_components/worx_vision_cloud/const.py
@@ -55,6 +55,7 @@ DEFAULT_CALENDAR_DAYS = 7
 BORDER_DISTANCE_OPTIONS_MM = (50, 100, 150, 200)
 
 SERVICE_START_ONE_TIME_MOWING = "start_one_time_mowing"
+SERVICE_START_ZONE_MOWING = "start_zone_mowing"
 SERVICE_SET_RTK_MAP_ID = "set_rtk_map_id"
 SERVICE_SET_FIRMWARE_NOTES = "set_firmware_notes"
 ATTR_VERSION = "version"
@@ -67,3 +68,4 @@ ATTR_MAP_ID = "map_id"
 ATTR_EDGE_CUT = "edge_cut"
 ATTR_RUNTIME = "runtime"
 ATTR_ZONES = "zones"
+ATTR_START_AT = "start_at"